+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -12,6 +12,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import nextstep.jdbc.exception.DataAccessException;
+import nextstep.jdbc.statementSetter.ArgumentPreparedStatementSetter;
 
 public abstract class JdbcTemplate {
